Example PR: Setup my IAM account
This is an example PR (do not merge!). It is a template to show
how to get the CI system to automatically provision a new IAM account.

Account creation workflow:

Create a PR similar to this one, add your your account to the
[account.yaml](https://github.com/Sage-Bionetworks/sandbox-infra/blob/master/templates/accounts.yaml)
file.  Make sure the AWSIAMXXXXXUser is a unique string.

* Have an AWS account admin review and approve the PR.
* Merge the PR
* The CI system will auto provision the account
* An email will be sent to the username with info setting up and using your
  new IAM account.

Removing an account:

AWS will refuse to remove users that have attached resources.
Therefore you must do the following before deleting them from this file:
* Detach or remove the following user resources: login profile, attached
  MFA device, access-keys, ssh-keys, and policies.
* Detach the user from all groups.

Needless to say, removing an account is an involved process.  Please contact the AWS
account admin for help.