setup a user, role and access keys to deploy PHCCP app
This will setup a specific user and a generic role for you to use to
deploy PHCCP. It will also generate the keys for your travis setup
to assume that role.  I'm not sure if the role matches what is
required for codedeploy, per code deploy instructions[1], but
it's something to get you started.

[1] https://docs.aws.amazon.com/codedeploy/latest/userguide/getting-started-create-service-role.html